chore(repo): basic refactoring and renaming
