+++ packages/gmail-api/index.ts
@@ -1,5 +1,10 @@
-export { batchModifyLabels, batchTrashThreads } from "./src/core-api";
-
-export { getAccessToken, listLabels, sendMail } from "./src/core-api";
+export {
+    batchModifyLabels,
+    batchTrashThreads,
+    getAccessToken,
+    listLabels,
+    sendMail
+} from "./src/core-api";
 export { incrementalSync } from "./src/routines/incremental-sync";
 export { partialSync } from "./src/routines/partial-sync";
+
