+++ packages/web-worker-logic/index.ts
@@ -1,4 +1,3 @@
 export { gmailApiWorker } from "./src/basic-client/gmail-api-worker";
-
 export { gmailBackgroundSyncWorker } from "./src/basic-client/gmail-bg-sync-worker";
 export { convertGmailEmailToClientDbEmail } from "./src/gmail-background-sync/lib/utils";
